feat: add admin components for data management and UI interactions
- Introduced ConfirmDialog component for confirmation prompts.
- Added DataTable component for displaying tabular data with pagination and empty state handling.
- Created EmptyState component to show a message when no data is available.
- Implemented FormModal component for modal forms with a close button.
- Developed IconButton component for clickable icons with various styles.
- Added Pagination component for navigating through paginated data.
- Created StatusBadge component for displaying status labels.

chore: add PowerShell scripts for database backup and restore

- Implemented backup-demo-data.ps1 for creating database and upload folder backups.
- Added backup-mysql.ps1 for backing up MySQL databases from Docker containers.
- Created export-local-db.ps1 for exporting local MySQL databases.
- Developed import-db-to-docker.ps1 for importing SQL files into Docker MySQL.
- Added restore-demo-data.ps1 for restoring data from backup packages.
- Implemented sync-upload-to-docker.ps1 for syncing local upload files to Docker.
- Created verify-data-integrity.ps1 for checking data consistency between local and Docker environments.
- Added utility functions in HanSportDataTools.ps1 for common operations related to Docker and file management.

+++ hansport_v2be/src/main/java/com/javaweb/domain/Order.java
@@ -18,7 +18,7 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private double totalPrice;
+    private long totalPrice;
 
     private String receiverName;
